The Beta Distribution Using Mean and Precision

Density, distribution function, quantile function, and random generation for the beta distribution with mean equal to mean and precision equal to prec (lower values correspond to higher variance).

dbetamp(x, mean, prec, log = FALSE)

pbetamp(q, mean, prec, lower.tail = TRUE, log.p = FALSE)

qbetamp(p, mean, prec, lower.tail = TRUE, log.p = FALSE)

rbetamp(n, mean, prec)

Arguments

x, q

Vector of quantiles (0 >= x,q <= 1).

mean

Vector of means for the beta distribution, where 0 > mean > 1.

prec

Vector of dispersion parameters for the beta distribution, where prec > 0.

log, log.p

Logical; if TRUE, probabilities p are given as log(p).

lower.tail

Logical; if TRUE (default), probabilities are P(X less than or equal to x) otherwise, P(X greater than x).

p

Vector of probabilities

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dbetamp gives the density, pbetamp gives the distribution function, qbetamp gives the quantile function, and rbetamp generates random deviates. The length of the result is determined by n for rbetamp, and is the maximum of the lengths of the numerical arguments for the other functions.

Details

The mean and precision for the beta distribution can be converted back to the standard positive shape parameters via shape1 = mean x prec and shape2 = (1-mean) x prec.

See stats::dbeta for more details.

Examples

x <- c(.25, .5, .75)
dbetamp( x, .5, 4)
#> [1] 1.125 1.500 1.125
# Same as 'dbeta' function
dbetamp( x, .5, 4) == dbeta( x, 2, 2)
#> [1] TRUE TRUE TRUE
pbetamp( x, .5, 4)
#> [1] 0.15625 0.50000 0.84375
qbetamp( pbetamp( x, .5, 4), .5, 4)
#> [1] 0.25 0.50 0.75
rbeta( 3, .5, 4)
#> [1] 0.09702619 0.05650044 0.25732298