Function to Left and Right-Censor Data

Function to left and right-censor numeric values in a column in a data frame.

Usage

ddabcd_data_censor(
  obj_data,
  chr_outcome = "ddabcd_indifference_score_1",
  lgc_indicator = FALSE,
  num_limits = c(0.78125, 99.21875),
  num_scaling = 100
)

Arguments

obj_data

Either a vector of values or a data frame.

chr_outcome

An optional character string, the column in obj_data to left and right-censor if obj_data is a data frame.

lgc_indicator

Logical; if TRUE returns an indicator for whether values were censored (-1 for left-censored, 1 for right-censored, and 0 otherwise) rather than the censored data itself.

num_limits

A numeric vector with the lower and upper limits below and above which to censor values, respectively. By default is set to the lowest and highest possible indifference scores found in the ABCD study data set.

num_scaling

A numeric value, a scaling constant by which to adjust the numeric values (useful, for instance, if data is in the range of 0 - 1 but the limits are in the range of 0 - 100).

Value

Either a numeric vector of censored value, or if lgc_indicator is TRUE, an integer vector with values of -1 for left-censored data, 1 for right-censored data, and 0 otherwise. If obj_data is NULL, returns the default limits used in the ABCD study.

Examples

# Example data set
dtf_data <-
  data.frame( ddabcd_indifference_score_1 = c( 0, 1, .5 ) )

# Indicator for censored values
ddabcd_data_censor( dtf_data, lgc_indicator = TRUE )
#> [1] -1  1  0
# Censored values
ddabcd_data_censor( dtf_data )
#> [1] 0.0078125 0.9921875 0.5000000

# Vector of values
ddabcd_data_censor(
  c(-5, 0, 5) %transform% 'logistic', chr_outcome = ''
)
#> [1] 0.0078125 0.5000000 0.9921875