Example Long-Form Data

An example data set with simulated indifference scores for 48 participants over two time points with separate rows for each delay duration, time point, and participant. Data are the result of the application of the ddabcd_data_prep function to the ddabcd_data_example_wide data set. Values were simulated from Rachlin's (2006) model with data following a logit-normal distribution. The generating log discounting rate and sensitivity parameters were specified to be weakly correlated with biological sex.

ddabcd_data_example_long

Format

ddabcd_data_example_long

A data frame with 672 rows and 21 columns:

src_subject_id

Participant ID

eventname

Study time point

age

Participant age in months

sex

Biological sex at birth for participant

ddis_scr_val_immedcho

Count of times participant picked rational immediate choice for 3 validity check questions

ddis_scr_val_immedcho

Count of times participant picked rational immediate choice for 3 validity check questions

ddabcd_simulation_lnk

The generating parameter value for the participant (log discounting rate)

ddabcd_simulation_s

The generating parameter value for the participant (sensitivity)

ddabcd_simulation_lnd

The generating parameter value for the participant (log standard deviation for logit-normal distribution)

ddacbd_quality_no_missing_data

Indicator of non-NA indifference scores

ddabcd_quality_picked_rational_immediate_choices

Indicator of when participant picked all 3 rational choices for validity check questions

ddabcd_quality_variation_in_responses

Indicator for when participant did not have the same indifference score across all delay durations

ddabcd_quality_met_Johson_and_Bickel_2008_criteria

Indicator for when participant's data met criteria of Johnson and Bickel (2008)

ddabcd_quality_significant_Mann_Kendall_test

Indicator for when participant's data had a significant one-sided Mann-Kendall test of monotonicity

ddabcd_delay_columns

The original column names for the indifference scores across different delay durations

ddabcd_indifference_score_100

Indifference scores ranging from 0 - 100

ddabcd_delay_months

Delay durations in months

ddabcd_delay_log_months

The log of the months for delay durations

ddabcd_delay_negative_log_months

The negative of the log of the months for delay durations

ddabcd_indifference_score_1

Indifference scores ranging from 0 - 1

ddabcd_indifference_score_log_odds

Indifference scores ranging from 0 - 1 converted to log-odds

ddabcd_indifference_score_censored

Indicator for when an indifference score (ranging from 0 - 100) was left-censored (value fixed to minimum of 0.78125) with a score of -1, right-censored (value fixed to maximum of 99.21875) with a score of 1, or not censored with a score of 0