Example Wide-Form Data in Format of ABCD Study

An example data set with simulated indifference scores for 48 participants over two time points formatted in the same style as data from the the ABCD study. Values were simulated from Rachlin's (2006) model with data following a logit-normal distribution. The generating log discounting rate and sensitivity parameters were specified to be weakly correlated with biological sex.

ddabcd_data_example_wide

Format

ddabcd_data_example_wide

A data frame with 96 rows and 15 columns:

src_subject_id

Participant ID

eventname

Study time point

age

Participant age in months

sex

Biological sex at birth for participant

ddis_scr_val_indif_pnt_1da

Indifference score for delay of 1 day

ddis_scr_val_indif_pnt_1week

Indifference score for delay of 1 day

ddis_scr_val_indif_pnt_1mth

Indifference score for delay of 1 month

ddis_scr_val_indif_pnt_3mth

Indifference score for delay of 3 months

ddis_scr_val_indif_pnt_1yr

Indifference score for delay of 1 year

ddis_scr_val_indif_pnt_5yr

Indifference score for delay of 5 years

ddis_scr_val_immedcho

Count of times participant picked rational immediate choice for 3 validity check questions

ddabcd_simulation_lnk

The generating parameter value for the participant (log discounting rate)

ddabcd_simulation_s

The generating parameter value for the participant (sensitivity)

ddabcd_simulation_lnd

The generating parameter value for the participant (log standard deviation for logit-normal distribution)