Compute Sum of Log Likelihoods for Delay Discounting Models

Function to compute the sum of the log-likelihoods for one of three delay discounting models: 1) an intercept-only model, 2) Mazur's (1987) model, or 3) Rachlin's (2006) model. Can be passed to optimization routines like stats::optim or bbmle::mle2 to conduct maximum likelihood estimation.

Usage

ddabcd_estimation_sum_of_log_likelihoods(
  num_param,
  dtf_data,
  chr_distribution,
  chr_measures = c("ddabcd_indifference_score_1", "ddabcd_delay_log_months"),
  num_fnscale = -1,
  lst_priors = NULL,
  fun_priors = NULL
)

Arguments

num_param

A numeric named vector of parameter values, where num_param['a0'] is the intercept only (logit-scale), num_param['lnk'] is for log discounting, num_param['s'] is for sensitivity, and num_param['lnd'] is log dispersion.

dtf_data

A data frame with the indifference scores (ranging from 0 to 1) and the log of the delay duration.

chr_distribution

A character string, the type of probability distribution the data are assumed to follow. Options include 'Normal', 'Beta', or 'Logit-normal'.

chr_measures

A character vector giving the column names in dtf_data for 1) the indifference scores, and 2) the log delays.

num_fnscale

An numeric value of either 1 or -1 adjusting the sum of the log-likelihoods (by default is -1 to ensure compatibility with functions like bbmle::mle2).

lst_priors

An optional named list with the hyperparameters for the priors on each parameter.

fun_prior

An optional function that takes the vector of parameters and a list of hyperparameters to compute a prior on the parameters. Allows for penalized maximum likelihood estimation (see ddabcd_estimation_priors).

Value

A numeric value, the sum of the log-likelihoods adjusted by num_fnscale.

Examples

# Example data set
data(ddabcd_data_example_long)
# Take data from 1st participant
dtf_delay <- ddabcd_data_example_long[1:7, ]

# Intercept-only model
# Data follow a normal distribution
num_param <- c( a0 = 2.2, lnd = log(.3) )
ddabcd_estimation_sum_of_log_likelihoods(
  num_param, dtf_delay, 'Normal'
)
#> [1] 0.3193902

# Mazur's (1987) model
# Data follow a beta distribution
num_param <- c( lnk = log(0.14), lnd = log(4) )
ddabcd_estimation_sum_of_log_likelihoods(
  num_param, dtf_delay, 'Beta'
)
#> [1] 6.658411

# Rachlin's (2006) model
# Data follow a logit-normal distribution
num_param <- c( lnk = log(0.14), s = 0.6, lnd = log(4) )
ddabcd_estimation_sum_of_log_likelihoods(
  num_param, dtf_delay, 'Logit-normal'
)
#> [1] 16.57022