Compute Sum of Log Tobit Likelihoods for Delay Discounting Models

Function to compute the sum of the log-likelihoods for one of three delay discounting models: 1) an intercept-only model, 2) Mazur's (1987) model, or 3) Rachlin's (2006) model. Can be passed to optimization routines like stats::optim or bbmle::mle2 to conduct maximum likelihood estimation. The outcome is required to be on the logit scale and is assumed to be both left and right-censored, and the tobit likelihood function (CITATION) is used as a result.

ddabcd_estimation_sum_of_log_tobit_likelihoods(
  num_param,
  dtf_data,
  chr_measures = c("ddabcd_indifference_score_log_odds",
    "ddabcd_delay_negative_log_months", "ddabcd_indifference_score_censored"),
  num_limits = NULL,
  num_fnscale = -1,
  lst_priors = NULL,
  fun_priors = NULL
)

Arguments

num_param

A numeric named vector of parameter values, where num_param['a0'] is the intercept only (logit-scale), num_param['nlnk'] is for minus log discounting, num_param['s'] is for sensitivity, and num_param['lnd'] is log dispersion.

dtf_data

A data frame with the log-odds of the indifference scores, the minus log of the delay duration, and an indicator for whether the data are left-censored (-1), right-censored (1), or uncensored (0).

chr_measures

A character vector giving the column names in dtf_data for 1) the log-odds of the indifference scores, 2) the minus log delays, and 3) the censoring indicator.

num_fnscale

An numeric value of either 1 or -1 adjusting the sum of the log-likelihoods (by default is -1 to ensure compatibility with functions like bbmle::mle2).

lst_priors

An optional named list with the hyperparameters for the priors on each parameter.

fun_prior

An optional function that takes the vector of parameters and a list of hyperparameters to compute a prior on the parameters. Allows for penalized maximum likelihood estimation.

Value

A numeric value, the sum of the log-tobit-likelihoods adjusted by num_fnscale.

Examples

# Examples