Fit Delay Discounting Models via Non-Linear Least Squares

Fits several delay discounting models to a long-form data set using R's built-in function stats::nls. This approach was originally used by Kohler et al. (2022) to fit Mazur's (1987) model to the delay discounting data from the ABCD study.

ddabcd_fit_models_using_nls(
  dtf_data,
  chr_outcome = "ddabcd_indifference_score_1",
  chr_log_delay = "ddabcd_delay_log_months",
  num_scaling = 1,
  lst_start = NULL,
  lgc_warnings = FALSE,
  lst_control = list(maxit = 20000),
  ...
)

Arguments

dtf_data

A data frame with separate rows for each delay duration and participant. Must have a column with the log of the delay durations.

chr_outcome

A character string, the column name for the indifference scores to fit.

chr_log_delay

A character string, the column name for the log of the delay durations.

num_scaling

A scaling constant (the numerator for the Mazur and Rachlin models). Typically the maximum possible indifference score.

lst_start

A named list consisting of the elements 'Null', 'M1987', and 'R2006'. Each list element should be a named vector with the starting values for the parameter estimates ('a0' for lst_start$Null, 'lnk' for lst_start$M1987, and 'lnk' and 's' for lst_start$R2006).

lgc_warnings

Logical; if TRUE displays a warning if any of the estimations failed.

lst_control

A list of control settings for stats::nls.

...

Additional parameters for stats::nls.

Value

A list with the stats::nls output for the null model, Mazur's (1987) model, and Rachlin's (2006) model.

References

Kohler, R. J., Lichenstein, S. D., & Yip, S. W. (2022). Hyperbolic discounting rates and risk for problematic alcohol use in youth enrolled in the Adolescent Brain and Cognitive Development study. Addiction Biology, 27 (2), 1-13. https://doi.org/10.1111/adb.13160

Mazur, J. E. (1987). An adjusting procedure for studying delayed reinforcement. In M. L. Commons, J. E. Mazur, J. A. Nevin, & H. Rachlin (Eds.), The effect of delay and intervening events on reinforcement value (pp. 55-73). Lawrence Erlbaum Associates, Inc.

Rachlin, H. (2006). Notes on discounting. Journal of the Experimental Analysis of Behavior, 85 (3), 425-435. https://doi.org/10.1901/jeab.2006.85-05

Examples

# Example data set
data(ddabcd_data_example_long)
# Take data from 1st participant
dtf_delay <- ddabcd_data_example_long[1:7, ]

# Fit simulated data for 1st participant in example data set
lst_fits.nls <- ddabcd_fit_models_using_nls(
  dtf_delay
)