Plot Delay Discounting Data (Single Panel)

Function to plot delay discounting data (typically the log of the delay durations in months and indifference scores as proportions) as a single panel figure.

Usage

ddabcd_plot_single_panel(
  obj_data,
  obj_x = "ddabcd_delay_log_months",
  obj_y = "ddabcd_indifference_score_1",
  chr_xlab = "log(Delay - months)",
  chr_ylab = "Indifference score (Proportion)",
  num_xlim = NULL,
  num_ylim = c(0, 1),
  num_dim = c(width = 5, height = 5),
  int_pch = 19,
  num_lwd = 2,
  num_cex.pt = 1,
  chr_color = c("black", "black", "blue", "grey"),
  num_jitter = 0,
  chr_model = "M1987",
  num_margin = c(4, 4, 2, 0.5),
  lst_predicted = NULL,
  lgc_add = FALSE,
  lgc_new = FALSE,
  lst_prediction_settings = NULL,
  ...
)

Arguments

obj_data

Either a data frame or a list of model fits (see for example ddabcd_fit_models_using_nls).

obj_x

Either a character vector with the column(s) to index for the x-axis values or a numeric vector.

obj_y

Either a character vector with the column(s) to index for the y-axis values or a numeric vector.

chr_xlab

A character string, the x-axis label.

chr_ylab

A character string, the y-axis label.

num_xlim

An optional numeric vector with the lower and upper limits for the x-axis.

num_ylim

A numeric vector with the lower and upper limits for the y-axis. Default is (0, 1).

num_dim

A numeric vector specifying the width and height of the plot panel when a new plotting window is generated.

int_pch

The type of point to draw.

num_lwd

A numeric value, the width of the line.

num_cex.pt

A numeric value, the size of the point.

chr_color

A character vector giving the colors to use for 1) the line for observed data, 2) the points for observed data, 3) the line for predicted data, and 4) the error bars for predicted data.

num_jitter

An optional numeric value specifying the limits for random uniform jitter noise to add to points.

chr_model

A character string, the model to extract results from if obj_data is a list of model fits. Options include 'Null', 'M1987', or 'R2006'.

num_margin

A numeric vector of 4 values giving the margin sizes for the bottom, left, top, and right edges of the figure.

lst_predicted

A list of either 1 or 3 elements specifying either the columns or numeric values to to use for [[1]] model predictions and [[2]] and [[3]] the lower and upper boundaries of the prediction interval.

lgc_add

Logical; if TRUE adds new data to an existing figure.

lgc_new

Logical; if TRUE generates a new plotting window.

lst_prediction_settings

An optional named list specifying how prediction intervals should be calculated if a model fit object is provided. The list should contain lst_prediction_settings$width for the width of the prediction interval, lst_prediction_settings$bootstrap set to TRUE for using a parametric bootstrap and FALSE otherwise, and lst_prediction_settings$samples with the number of samples to draw for the parametric bootstrap.

...

Additional parameters to the base::plot function.

Value

A plot.

Examples

if (FALSE) {
# Load in example data sets
data( ddabcd_data_example_wide )
data( ddabcd_data_example_long )
dtf_wide <- ddabcd_data_example_wide %index% list( 1, 'eventname' )
dtf_long <- ddabcd_data_example_long %index% list( 1, 'eventname' )

# By default data assumed to be long-form
ddabcd_plot_single_panel( dtf_long %index% 1 ) # 1st participant
# Wide-form data can be plotted by providing vector of column names
ddabcd_plot_single_panel(
  dtf_wide %index% 1, # 1st participant
  obj_x = ddabcd_data_delay_durations(),
  colnames( dtf_wide %col% 'indif' ),
  num_ylim = c( 0, 100 )
)

# Example using model fits for first participant
lst_fits.nls <- ddabcd_fit_models_using_nls( dtf_long %index% 1)
ddabcd_plot_single_panel( lst_fits.nls, chr_model = 'M1987' )
# Add predictions for Rachlin's (2006) model
ddabcd_plot_single_panel(
  lst_fits.nls, chr_model = 'R2006',
  chr_color = c('black', 'black', 'red' ), lgc_add = TRUE
)
}