Extract 'ddabcd_attr' Component

Utility function to extract elements from the 'ddabcd_attr' component of the attributes for model fits from ddabcd_fit_models_using_nls and ddabcd_fit_models_using_mle2.

Usage

ddabcd_post_fit_attr(obj_fit, chr_attr = "", obj_default = NA)

Arguments

obj_fit

An R object for model fit results.

chr_attr

A character string, the specific attribute to access. Options include 'predicted', 'residuals', 'rmse', 'loglikelihood', 'AIC', 'BIC', 'nobs', or 'dof'.

obj_default

Value to return if attributes aren't found.

Value

A list of elements or a specific element from the 'ddabcd_attr' attribute.

Examples

# Example data
data(ddabcd_data_example_long)
# Use data for first time point and first 3 participants
dtf_long <- ddabcd_data_example_long %index% 1:3
dtf_long <- dtf_long %index% list( 1, 'eventname' )

# Fit data using 'nls'
lst_all_fits.nls <- ddabcd_fit_models_across_participants(
  dtf_long, ddabcd_fit_models_using_nls, lgc_progress = FALSE
)

ddabcd_post_fit_attr( lst_all_fits.nls[[1]]$M1987 )['rmse']
#> $rmse
#> [1] 0.1649162
#>