Bootstrapped Prediction Intervals for Models fit by mle2

Compute bootstrapped prediction intervals via simulating

possible parameter values from a multivariate normal distribution, and 2) possible observed data values for delay discounting model at each set of parameter values.

ddabcd_post_fit_bootstrapped_prediction_intervals_mle2(
  lst_fit,
  int_samples = 1000,
  lgc_suppress_error = FALSE,
  chr_labels = c("a0", "lnk", "s", "lnd"),
  num_width = 0.95,
  num_limits = NULL,
  ...
)

Arguments

lst_fit

A model fit object that has methods for coef and vcov.

int_samples

The numer of samples to draw from the multivariate normal distribution.

lgc_suppress_error

Logical; if TRUE suppress error messages and return a matrix of NA values instead.

chr_labels

The column names for the matrix of NA values if samples cannot be drawn.

num_width

A numeric value between 0 and 1, the width of the prediction interval.

num_limits

An optional vector giving the lower and upper limits below and above which data should be censored.

Value

A 2 x N matrix whose rows give the lower and upper bootstrapped prediction intervals for the N observations.

Examples

# Example data
data( ddabcd_data_example_long )
dtf_long <- ddabcd_data_example_long %index% 1
dtf_long <- dtf_long %index% list( 1, 'eventname' )

lst_fits.mle2 <- ddabcd_fit_models_using_mle2(
  dtf_long, chr_distribution = 'Logit-normal'
)

ddabcd_post_fit_bootstrapped_prediction_intervals_mle2(
  lst_fits.mle2$R2006
)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#> 2.5%  0.7873843 0.7168757 0.6026481 0.5242532 0.4188485 0.3229774 0.2105051
#> 97.5% 0.9750150 0.9629716 0.9297762 0.9002964 0.8625759 0.8134035 0.7647308