Index Rows of a Data Frame

Convenience operator that matches a user-supplied index (either an integer or character string) against the elements in a data frame column and returns the corresponding set of rows. Useful, for example, to isolate rows for a specific participant in a long-form data set.

dtf_data %index% obj_index

Arguments

dtf_data

A data frame.

obj_index

Either an integer vector, character vector, or a list with the integer/character vector and a character string specifying the column to match over. If obj_index is an integer, the operator uses the first column of dtf_data to match over.

Value

A data frame, the subset of dtf_data

that had successful matches.

Examples

# Create example data set
dtf_data <- data.frame(
  V1 = rep( LETTERS[1:3], each = 2 ), V2 = c( 1:3, 1:3 )
)

# Return rows for first unique element of 'V1'
dtf_data %index% 1
#>   V1 V2
#> 1  A  1
#> 2  A  2
# Return rows for first unique element of 'V2'
dtf_data %index% list( 1, 'V2' )
#>   V1 V2
#> 1  A  1
#> 4  B  1
# Return rows for 'V1' == 'B'
dtf_data %index% 'B'
#>   V1 V2
#> 3  B  3
#> 4  B  1
dtf_data %index% list( '3', 'V2' )
#>   V1 V2
#> 3  B  3
#> 6  C  3