Compute Priors Over Delay Discounting Model Parameters

Function to compute priors over delay discounting model parameters.

Usage

ddabcd_estimation_priors(num_param, lst_priors, chr_distribution)

Arguments

num_param

A numeric named vector of parameter values, where num_param['a0'] is the intercept only (logit-scale), num_param['lnk'] is for log discounting, num_param['s'] is for sensitivity, and num_param['lnd'] is log dispersion.

lst_priors

A named list of vectors specifying the mean and standard deviation, respectively, for the prior on each parameter.

chr_distribution

A character string specifying the distribution for the data. Options are 'Normal', 'Beta', or 'Logit-normal'.

Value

A numeric value, a sum of log-likelihoods.

Details

The following parameters are given normal priors with user-specified means and standard deviations:

"a0"intercept only on logit-scale

"lnk"log discounting

"s"sensitivity

"lnd"log dispersion

The interpretation of the log dispersion parameter depends on the distribution for the data. For normal and logit-normal distributions lower values indicate lower variance, whereas for the beta distribution, higher values indicate lower variance.

Examples

# Example normal prior with mean of 0 and
# sd of 1 on intercept on logit-scale
ddabcd_estimation_priors(
  c( a0 = 0 ), lst_priors = list( a0 = c( 0, 1 ) )
)
#> [1] -0.9189385