Predictions and Simulations From Delay Discounting Models

Generates predicted or simulated indifference scores for different delay discounting models. Supported models are

a intercept-only model, 2) Mazur's (1987) model, and

Rachlin's (2006) model.

ddabcd_model_predictions(
  num_param,
  num_delay,
  chr_distribution = "Normal",
  lgc_logit = FALSE,
  num_scaling = 1
)

Arguments

num_param

A named numeric vector of parameter values. Parameter names can be b0 for the intercept-only model, a0 for the intercept-only model (logit-scale), k for the discounting parameter, lnk for the log discounting parameter, nlnk for the negative log discounting parameter, s for the sensitivity parameter (Rachlin, 2006), d for the dispersion parameter, and lnd for the log of the dispersion parameter. If d or lnd are provided, function generates random deviates from the model, otherwise it produces model predictions. The interpretation of d and lnd depend on the specification of chr_error.

num_delay

A numeric vector of delay durations (for num_param['k']), log delay durations (for num_param['lnk']), or minus log durations (for num_param['nlnk']).

lgc_logit

Logical; if TRUE

num_scaling

A numeric value, rescales results if lgc_logit is FALSE.

chr_error

The distribution for the dispersion around model predictions. Options are 'Normal', 'Beta', or 'Logit-normal'.

Value

A vector of predicted or simulated indifference scores.

Details

For an indifference score I scaled to be between (0 - 1), Mazur's (1987) model is I = 1/(1 + kdelay). Rachlin's (2006) model is I = 1/(1 + k(delay^s)). Mazur's model therefore is a special case of Rachlin's when s = 1. The models can be rewritten as a linear equation on the logit-scale: logit(I) = -ln(k) - ln(delay)*s.

References

Mazur, J. E. (1987). An adjusting procedure for studying delayed reinforcement. In M. L. Commons, J. E. Mazur, J. A. Nevin, & H. Rachlin (Eds.), The effect of delay and intervening events on reinforcement value (pp. 55-73). Lawrence Erlbaum Associates, Inc.

Rachlin, H. (2006). Notes on discounting. Journal of the Experimental Analysis of Behavior, 85 (3), 425-435. https://doi.org/10.1901/jeab.2006.85-05

Examples

# Delay duration in months
x <- c( 1, 3, 12 )

# Mazur's (1987) model
k <- 0.14
# Predictions given different parameterizations
ddabcd_model_predictions( c( k = k ), x )
#> [1] 0.8771930 0.7042254 0.3731343
ddabcd_model_predictions( c( lnk = log(k) ), log(x) )
#> [1] 0.8771930 0.7042254 0.3731343
ddabcd_model_predictions( c( nlnk = -log(k) ), -log(x) )
#> [1] 0.8771930 0.7042254 0.3731343

# Rachlin's (2006) model
s = .60
ddabcd_model_predictions( c( k = k, s = s ), x )
#> [1] 0.8771930 0.7870016 0.6166064
ddabcd_model_predictions( c( lnk = log(k), s = s ), log(x) )
#> [1] 0.8771930 0.7870016 0.6166064
ddabcd_model_predictions( c( nlnk = -log(k), s = s ), -log(x) )
#> [1] 0.8771930 0.7870016 0.6166064

# Random deviates
# Normal model
ddabcd_model_predictions(
  c( k = k, d = .3 ), x, chr_distribution = 'Normal'
)
#> [1]  0.9275767339  1.1458483960 -0.0002954844
# Beta model
ddabcd_model_predictions(
  c( k = k, d = 4 ), x, chr_distribution = 'Beta'
)
#> [1] 0.99448076 0.71440649 0.06532193
# Logit-normal model
ddabcd_model_predictions(
  c( k = k, d = 0.3 ), x, chr_distribution = 'Logit-normal'
)
#> [1] 0.8921436 0.6729966 0.3501813