Compute Prediction Intervals

Compute prediction intervals for delayed discounting model fit. Users can either compute the interval using model parameter point estimates, or incorporate uncertainty in parameter estimates in the prediction interval via a parametric bootstrap approach.

Usage

ddabcd_post_fit_prediction_intervals(
  obj_fit,
  num_width = 0.95,
  lgc_bootstrap = FALSE,
  int_samples = 1000,
  num_limits = NULL,
  lgc_suppress_error = FALSE,
  ...
)

Arguments

obj_fit

An R object for model fit results.

num_width

A numeric value between 0 and 1 specifying the width of the prediction interval.

lgc_bootstrap

Logical; if TRUE computes prediction intervals via a parametric bootstrap approach that incorporates uncertainty in model parameters.

int_samples

An integer value, the number of samples to use for the parameter bootstrap (higher values lead to less Monte Carlo error of approximation but longer computation times).

num_limits

An optional numeric vector giving the lower and upper limits below and above which simulated predictions should be censored.

lgc_suppress_error

Logical; if TRUE function does not produce an error when the parameter bootstrap cannot be computed.

...

Additional parameters for the ddabcd_model_predictions function.

Value

A 2xN matrix whose rows give the lower and upper prediction limits for the N data points, respectively.

Examples

# Example data
data("ddabcd_data_example_long")
dtf_long <- ddabcd_data_example_long %index% 1
dtf_long <- dtf_long %index% list( 1, 'eventname' )

lst_fits.nls <- ddabcd_fit_models_using_nls( dtf_long )
lst_fits.mle2 <- ddabcd_fit_models_using_mle2(
  dtf_long, chr_distribution = 'Logit-normal'
)

# Prediction intervals using point estimates
ddabcd_post_fit_prediction_intervals( lst_fits.nls$M1987 ) |> round(2)
#>      [,1] [,2] [,3] [,4] [,5] [,6]  [,7]
#> [1,] 0.65 0.65 0.64 0.59 0.48 0.20 -0.16
#> [2,] 1.35 1.35 1.33 1.28 1.18 0.89  0.54
ddabcd_post_fit_prediction_intervals( lst_fits.mle2$M1987 ) |> round(2)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,] 0.04 0.04 0.04 0.03 0.02 0.02 0.02
#> [2,] 0.99 0.99 0.99 0.99 0.99 0.99 0.99

# Bootstrapped prediction intervals
ddabcd_post_fit_prediction_intervals(
  lst_fits.mle2$M1987,
  lgc_bootstrap = TRUE
) |> round(2)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> 2.5%  0.76 0.41 0.12 0.01 0.01 0.00 0.00
#> 97.5% 1.00 1.00 1.00 1.00 0.99 0.95 0.78